fixed crash issue and code clean up
