+++ app/src/main/java/aaa/app/android/sqlroomsample/util/APPConstant.kt
@@ -3,4 +3,5 @@ package aaa.app.android.sqlroomsample.util
 object APPConstant {
 
     const val DATE_FORMAT_ONE="dd-MM-yyyy"
+    const val TIME_FORMAT_ONE="hh:mm a"
 }
