--- app/src/main/java/aaa/app/android/sqlroomsample/util/FragmentExt.kt
@@ -1,64 +1,0 @@
-package aaa.app.android.sqlroomsample.util
-
-import android.app.DatePickerDialog
-import android.app.TimePickerDialog
-import android.content.Context
-import com.google.android.material.textfield.TextInputEditText
-import java.text.SimpleDateFormat
-import java.util.*
-
-fun TextInputEditText.transformIntoDatePicker(context: Context, format: String, maxDate: Date? = null) {
-    isFocusableInTouchMode = false
-    isClickable = true
-    isFocusable = false
-    isCursorVisible = false
-
-    val myCalendar = Calendar.getInstance()
-    val datePickerOnDataSetListener =
-        DatePickerDialog.OnDateSetListener { _, year, monthOfYear, dayOfMonth ->
-            myCalendar.set(Calendar.YEAR, year)
-            myCalendar.set(Calendar.MONTH, monthOfYear)
-            myCalendar.set(Calendar.DAY_OF_MONTH, dayOfMonth)
-            val sdf = SimpleDateFormat(format, Locale.getDefault())
-            setText(sdf.format(myCalendar.time))
-        }
-
-    setOnClickListener {
-        DatePickerDialog(
-            context, datePickerOnDataSetListener, myCalendar
-                .get(Calendar.YEAR), myCalendar.get(Calendar.MONTH),
-            myCalendar.get(Calendar.DAY_OF_MONTH)
-        ).run {
-            maxDate?.time?.also { datePicker.maxDate = it }
-            show()
-        }
-    }
-}
-
-
-fun TextInputEditText.transformIntoTimePicker(context: Context, format: String, maxDate: Date? = null) {
-    isFocusableInTouchMode = false
-    isClickable = true
-    isFocusable = false
-    isCursorVisible = false
-
-    val myCalendar = Calendar.getInstance()
-
-    val timePickerOnDataSetListener = TimePickerDialog.OnTimeSetListener { _, hour, minute ->
-        myCalendar.set(Calendar.HOUR_OF_DAY, hour)
-        myCalendar.set(Calendar.MINUTE, minute)
-
-        setText(SimpleDateFormat(format, Locale.getDefault()).format(myCalendar.time))
-    }
-
-
-    setOnClickListener {
-        TimePickerDialog(
-            context,
-            timePickerOnDataSetListener,
-            myCalendar.get(Calendar.HOUR_OF_DAY),
-            myCalendar.get(Calendar.MINUTE),
-            true
-        ).show()
-    }
-}
